refactor: consolidate auth tests and extract idle timer logic
Remove duplicate rate limiting tests from auth.test.ts, standardize
imports to use @/ alias, and extract idle timer into reusable hook
