+++ src/renderer/components/Sidebar.tsx
@@ -11,7 +11,7 @@ import {
 } from 'lucide-react';
 import { useTranslation } from '@/i18n';
 import { useNavigation, Page } from '@/renderer/store';
-import { useAuthStore } from '@/renderer/store/useAuthStore';
+import { useAuthStore } from '@/renderer/store';
 import SidebarButton from './SidebarButton';
 import { Logo } from '.';
