+++ src/renderer/hooks/index.ts
@@ -3,3 +3,4 @@ export { useTestEvents } from './useTestEvents';
 export { useTestInput } from './useTestInput';
 export { useAttentionMetrics } from './useAttentionMetrics';
 export { useAuthGuard, type AuthModalStatus } from './useAuthGuard';
+export { useIdleTimer } from './useIdleTimer';
